+++ elide-async/src/main/java/com/yahoo/elide/async/hooks/AsyncQueryHook.java
@@ -8,19 +8,20 @@ package com.yahoo.elide.async.hooks;
 import com.yahoo.elide.annotation.LifeCycleHookBinding.Operation;
 import com.yahoo.elide.annotation.LifeCycleHookBinding.TransactionPhase;
 import com.yahoo.elide.async.models.AsyncAPI;
+import com.yahoo.elide.async.models.AsyncAPIResult;
 import com.yahoo.elide.async.models.AsyncQuery;
 import com.yahoo.elide.async.models.QueryType;
 import com.yahoo.elide.async.operation.AsyncAPIOperation;
 import com.yahoo.elide.async.operation.GraphQLAsyncQueryOperation;
 import com.yahoo.elide.async.operation.JSONAPIAsyncQueryOperation;
 import com.yahoo.elide.async.service.AsyncExecutorService;
-import com.yahoo.elide.async.service.thread.AsyncAPICallable;
 import com.yahoo.elide.core.exceptions.InvalidOperationException;
 import com.yahoo.elide.core.security.ChangeSpec;
 import com.yahoo.elide.core.security.RequestScope;
 import com.yahoo.elide.graphql.QueryRunner;
 
 import java.util.Optional;
+import java.util.concurrent.Callable;
 
 /**
  * LifeCycle Hook for execution of AsyncQuery.
@@ -34,9 +35,8 @@ public class AsyncQueryHook extends AsyncAPIHook<AsyncQuery> {
     @Override
     public void execute(Operation operation, TransactionPhase phase, AsyncQuery query, RequestScope requestScope,
             Optional<ChangeSpec> changes) {
-        AsyncAPICallable queryWorker = new AsyncAPICallable(query, getOperation(query, requestScope),
-                (com.yahoo.elide.core.RequestScope) requestScope);
-        executeHook(operation, phase, query, requestScope, queryWorker);
+        Callable<AsyncAPIResult> callable = getOperation(query, requestScope);
+        executeHook(operation, phase, query, requestScope, callable);
     }
 
     @Override
@@ -55,9 +55,11 @@ public class AsyncQueryHook extends AsyncAPIHook<AsyncQuery> {
     public AsyncAPIOperation<?> getOperation(AsyncAPI query, RequestScope requestScope) {
         AsyncAPIOperation<?> operation = null;
         if (query.getQueryType().equals(QueryType.JSONAPI_V1_0)) {
-            operation = new JSONAPIAsyncQueryOperation(getAsyncExecutorService());
+            operation = new JSONAPIAsyncQueryOperation(getAsyncExecutorService(), query,
+                    (com.yahoo.elide.core.RequestScope) requestScope);
         } else {
-            operation = new GraphQLAsyncQueryOperation(getAsyncExecutorService());
+            operation = new GraphQLAsyncQueryOperation(getAsyncExecutorService(), query,
+                    (com.yahoo.elide.core.RequestScope) requestScope);
         }
         return operation;
     }
