+++ elide-async/src/main/java/com/yahoo/elide/async/operation/AsyncAPIOperation.java
@@ -7,19 +7,12 @@ package com.yahoo.elide.async.operation;
 
 import com.yahoo.elide.async.models.AsyncAPI;
 import com.yahoo.elide.async.models.AsyncAPIResult;
-import com.yahoo.elide.core.RequestScope;
+
+import java.util.concurrent.Callable;
 
 /**
  * AsyncAPI Execute Operation Interface.
  * @param <T> Type of AsyncAPI.
  */
-public interface AsyncAPIOperation<T extends AsyncAPI> {
-
-    /**
-     * Execute the AsyncAPI request.
-     * @param queryObj AsyncAPI type object.
-     * @param scope RequestScope object.
-     * @return AsyncAPIResult object
-     */
-    public abstract AsyncAPIResult execute(AsyncAPI queryObj, RequestScope scope);
+public interface AsyncAPIOperation<T extends AsyncAPI> extends Callable<AsyncAPIResult> {
 }
